refactor: get_constructions
