+++ energetica/routers/__init__.py
@@ -22,6 +22,7 @@ from .daily_quiz import router as daily_quiz_router
 from .map import router as map_router
 from .networks import router as network_router
 from .players import router as player_router
+from .projects import router as projects_router
 from .resource_market import router as resource_market_router
 from .scoreboard import router as scoreboard_router
 from .shipments import router as shipments_router
@@ -37,6 +38,7 @@ api_routers = [
     map_router,
     network_router,
     player_router,
+    projects_router,
     resource_market_router,
     shipments_router,
     scoreboard_router,
